Split grammar: one rule file per entity.
